refactor(backend): extract sync notification handling
